2022/05/03 1. JPA ENTITY 추가 2. SwaggerConfig 설정 3. DatabaseConfig 설정 4. Test Data 주입 > 에러발생

+++ src/main/java/com/JS_Blog/springBlog/blog/entity/Tags.java
@@ -32,7 +32,7 @@ public class Tags {
       @GeneratedValue(strategy = GenerationType.IDENTITY)
       @Id
       @Column(name = "tag_no")
-      private short tagNo;
+      private int tagNo;
       @Basic
       @Column(name = "tag_name")
       private String tagName;
